\name{table36}

\alias{table36}

\docType{data}

\title{Beijing development zone data}

\description{
independent variable x1: Cumulative number of development zones
independent variable x2: China Merchants registered capital 
dependent variable y : Sales revenue in the Development Zone 

}

\usage{data(table36)}

\examples{
data(table36)
}

