\name{table51}

\alias{table51}

\docType{data}

\title{Sales of some consumer goods}

\description{
independent variable x1: Disposable income of residents
ndependent variable x2: Price index of the consumer goods
independent variable x3: Average price index of other consumer goods
dependent variable y : Sales of some consumer goods
}

\usage{data(table51)}

\examples{
data(table51)
}

