\name{tableA17}

\alias{tableA17}

\docType{data}

\title{The proportion of urban and rural residents in the proportion of regular savings }

\description{The proportion of urban and rural residents in the proportion of regular savings in BeiJing
}

\usage{data(tableA17)}

\examples{
data(tableA17)
}

