\name{zhbit6-package}
\alias{zhbit6-package}
\alias{zhbit7}
\docType{package}
\title{
\packageTitle{zhbit7}
}
\description{
\packageDescription{zhbit7}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{zhbit7}
\packageIndices{zhbit7}
}
\author{
\packageAuthor{zhbit7}
Maintainer: \packageMaintainer{zhbit7}
}
\references{
regression
}

\keyword{ package }
\seealso{
The model fitting function \code{\link{lm}}
}
